Switch to 16 bit ints to reduce channel data copy time

+++ Beamformer/Beamformer/ChannelData.swift
@@ -4,7 +4,7 @@ import Foundation
 
 public struct ChannelData
 {
-    public var complexSamples: [Float]
+    public var complexSamples: [Int16]
     public var numberOfChannels: Int
     public var numberOfSamplesPerChannel: Int
 }
